cast every sensor value to a double

+++ lib/src/models/devices/open_earable_factory.dart
@@ -300,7 +300,7 @@ class _OpenEarableSensor extends Sensor {
           continue;
         }
 
-        values.add(entry.value);
+        values.add(entry.value.toDouble());
       }
 
       SensorValue sensorValue = SensorValue(
